Move dragdrop to src root
