+++ lib/src/main/java/com/zetcode/dragdrop/ListDnD.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.dragdrop;
 
 import javax.swing.DefaultListModel;
 import javax.swing.DropMode;
