+++ lib/src/main/java/com/zetcode/dragdrop/plain.png
