+++ lib/src/main/java/com/zetcode/dragdrop/sad.png
