+++ lib/src/main/java/com/zetcode/dragdrop/smile.png
